backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic>=2.0.0
python-jose[cryptography]
passlib[argon2]
argon2-cffi
razorpay
requests
cloudinary
python-multipart
python-dotenv
python-multipart
psycopg2-binary
supabase
